features: float32 throughout; TODO: streaming refactor for OOM ceiling
The lag-windowed features tensor was being built in float64 with no
accuracy benefit (downstream JAX trainer casts to float32 anyway).
2x memory tax on every loaded ticker. At 19 tickers × ~3000 valid
bars × K=96 × F=33 that's the difference between 720 MB and 1.4 GB
on host RAM, which combined with JAX device copies + FiLM auxiliary
arrays + Colab CPU's 12 GB limit was OOM-killing the supervised
sign-of-return run before training started (2026-05-01).

Fix: cast feats to float32 at construction in build_lagged_features,
and stop the upcasts in build_features_and_targets (compute_scalogram
already returns float32; the .astype(np.float64) calls were upcasting
back to no benefit). Also cast the optional channels (mu/std/return/
return-sign) to float32 for consistency.

Halves the static feature memory; should fit the 19-ticker / K=96 /
full-FiLM-grid configuration on a 12 GB Colab CPU runtime. Doesn't
change asymptote — at K > 128 or pool > 30 tickers we'll OOM again.

Adds TODO section "Streaming feature pipeline" sketching the np.memmap
+ batch-sampler refactor that would eliminate the bulk pre-compute
entirely (target: peak RSS stays under 4 GB at 100 tickers / K=128).

+++ apps/notebook/src/ss_notebook/replay/features.py
@@ -181,7 +181,7 @@ def build_lagged_features(
     if window_cols < 1:
         raise ValueError(f'window_cols must be >= 1, got {window_cols}')
     C, n_dates = channels_cn.shape
-    feats = np.full((n_dates, window_cols, C), np.nan, dtype=np.float64)
+    feats = np.full((n_dates, window_cols, C), np.nan, dtype=np.float32)
     for k in range(window_cols):
         feats[k:, k] = channels_cn[:, :n_dates - k].T
     return feats.reshape(n_dates, window_cols * C)
@@ -230,18 +230,20 @@ def build_features_and_targets(
         raise ValueError('include_returns and include_return_sign are '
                          'mutually exclusive — they share a channel slot.')
     coeffs, power = compute_scalogram(prices, scales, lookback=lookback)
+    # All channels kept in float32 — float64 was a 2x memory tax with no
+    # accuracy benefit (downstream JAX trainer casts to float32 anyway).
     channels: list[np.ndarray] = [
-        coeffs.astype(np.float64),
-        power.astype(np.float64),
+        coeffs.astype(np.float32),
+        power.astype(np.float32),
     ]
     if include_zscore_stats:
         mu, std = rolling_zscore_stats(prices, lookback=lookback)
-        channels.append(mu[None, :])
-        channels.append(std[None, :])
+        channels.append(mu[None, :].astype(np.float32))
+        channels.append(std[None, :].astype(np.float32))
     if include_returns:
-        channels.append(log_returns(prices)[None, :])
+        channels.append(log_returns(prices)[None, :].astype(np.float32))
     elif include_return_sign:
-        channels.append(log_return_signs(prices)[None, :])
+        channels.append(log_return_signs(prices)[None, :].astype(np.float32))
     channels_cn = np.vstack(channels)
     features = build_lagged_features(channels_cn, window_cols)
